Mob 1705 default browser experiment (#414)
* wip

* use Intro flag for display logic of default promo

* tweak logic, fix appcenter

* Add Staging config to Beta

* fix unit tests

* some cleanup

* remove branch from app center build

* add creating config to initial bootstrap

* Use XCode 14.2.0

* use main branch / update l10n

* Harden flaky test by increasing timeout
