+++ .circleci/config.yml
@@ -5,7 +5,7 @@ jobs:
     environment:
       CI: true
     macos:
-      xcode: "14.1.0"
+      xcode: "14.2.0"
     resource_class: macos.x86.medium.gen2
     steps:
       - checkout
@@ -33,7 +33,7 @@ jobs:
     environment:
       CI: true
     macos:
-      xcode: "14.1.0"
+      xcode: "14.2.0"
     resource_class: macos.x86.medium.gen2
     steps:
       - checkout
@@ -61,7 +61,7 @@ jobs:
     environment:
       CI: true
     macos:
-      xcode: "14.1.0"
+      xcode: "14.2.0"
     resource_class: macos.x86.medium.gen2
     steps:
       - checkout
@@ -75,6 +75,9 @@ jobs:
       - run:
           name: verify github
           command: for ip in $(dig @8.8.8.8 github.com +short); do ssh-keyscan github.com,$ip; ssh-keyscan $ip; done 2>/dev/null >> ~/.ssh/known_hosts
+      - run:
+          name: extract Staging XCConfig
+          command: echo "$STAGING_XCCONFIG" | base64 --decode > Client/Configuration/Staging.xcconfig
       - run:
           name: clean packages
           command: xcodebuild -scheme Ecosia -project Client.xcodeproj clean
