+++ .github/workflows/UnitTests.yml
@@ -10,7 +10,7 @@ jobs:
   test:
     runs-on: macos-12
     env:
-      DEVELOPER_DIR: /Applications/Xcode_14.1.0.app/Contents/Developer
+      DEVELOPER_DIR: /Applications/Xcode_14.2.0.app/Contents/Developer
 
     steps:
     - uses: actions/checkout@v2
@@ -28,6 +28,8 @@ jobs:
       run: find . -type f \( -name '*.pbxproj' -o -name 'Package.swift' -o -name 'Package.resolved' \) -exec sed -i '' "s/https:\/\/github.com\/${GITHUB_REPOSITORY_OWNER}/https:\/\/${GITHUB_ACCESS_TOKEN}@github.com\/${GITHUB_REPOSITORY_OWNER}/g" {} \;
       env:
         GITHUB_ACCESS_TOKEN: ${{ secrets.IOS_CORE_TOKEN }}
+    - name: Create empty Staging xcconfig
+      run: touch Client/Configuration/Staging.xcconfig
     - name: clean packages
       run: xcodebuild -scheme Ecosia -project Client.xcodeproj clean
     - name: run unit tests
