+++ Client.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -59,7 +59,7 @@
       "location" : "https://github.com/ecosia/ios-core.git",
       "state" : {
         "branch" : "main",
-        "revision" : "8eccaab373f6ed4bd2a2c22b1f6d51f0fa0727a9"
+        "revision" : "de43cdeee37c3094427a18ac1422db6be6055104"
       }
     },
     {
