+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -938,6 +938,12 @@ class BrowserViewController: UIViewController {
 
         // Make sure reload button is hidden on homepage
         urlBar.locationView.reloadButton.reloadButtonState = .disabled
+
+        DispatchQueue.main.async {
+            if !User.shared.firstTime {
+                self.presentDefaultBrowserPromoIfNeeded()
+            }
+        }
     }
 
     /// Once the homepage is created, browserViewController keeps a reference to it, never setting it to nil during
@@ -2221,17 +2227,10 @@ extension BrowserViewController: UIAdaptivePresentationControllerDelegate {
 
 extension BrowserViewController {
     func presentIntroViewController(_ alwaysShow: Bool = false) {
-        if (User.shared.migrated != true && !User.shared.firstTime)
-            || User.shared.referrals.pendingClaim != nil {
+        if showLoadingScreen(for: .shared) {
             present(LoadingScreen(profile: profile, referrals: referrals, referralCode: User.shared.referrals.pendingClaim), animated: true)
         } else if User.shared.firstTime {
-            if #available(iOS 14, *) {
-                let defaultPromo = DefaultBrowser(delegate: self)
-                present(defaultPromo, animated: true)
-            } else {
-                User.shared.firstTime = false
-            }
-            profile.prefs.setInt(1, forKey: PrefsKeys.IntroSeen)
+            User.shared.firstTime = false
             User.shared.migrated = true
             User.shared.hideRebrandIntro()
             // deactivate searchbar hint for new users
@@ -2242,6 +2241,26 @@ extension BrowserViewController {
             intro.modalTransitionStyle = .crossDissolve
             present(intro, animated: true)
             User.shared.hideRebrandIntro()
+        }
+        presentDefaultBrowserPromoIfNeeded()
+    }
+
+    private func showLoadingScreen(for user: User) -> Bool {
+        (user.migrated != true && !user.firstTime)
+            || user.referrals.pendingClaim != nil
+    }
+
+    func presentDefaultBrowserPromoIfNeeded() {
+        guard !showLoadingScreen(for: .shared),
+              !User.shared.showsRebrandIntro else { return }
+
+        if shouldShowIntroScreen && Unleash.getRequiredSearches() <= User.shared.treeCount  {
+            if #available(iOS 14, *) {
+                let defaultPromo = DefaultBrowser(delegate: self)
+                present(defaultPromo, animated: true)
+            } else {
+                profile.prefs.setInt(1, forKey: PrefsKeys.IntroSeen)
+            }
         }
     }
